Refactor ReadySection
