+++ apps/nameguard.io/src/components/molecules/index.ts
@@ -3,3 +3,4 @@ export * from "./NGReport";
 export * from "./NGSearch";
 export * from "./NameGuardWrapper";
 export * from "./DeveloperResourceItem";
+export * from "./CodeSnippet";
