+++ apps/nameguard.io/src/components/organisms/ReadySection.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import cc from "classcat";
-import { highlight } from "sugar-high";
 import { ImageCharacteristics } from "@/types/imageTypes";
+import { CodeSnippet } from "@components/molecules";
 
 export type ReadySectionProps = {
   sectionTargetClientMessage: string;
@@ -16,38 +16,6 @@ export type ReadySectionProps = {
 };
 
 export function ReadySection(props: ReadySectionProps) {
-  const mediaDiv =
-    "flex sm:flex-col xl:flex-row justify-center border-0 rounded-none items-center gap-12 w-full h-full xl:h-3/4 py-16 px-10 flex-shrink-0 sm:gap-10";
-  const backgroundDiv = cc([
-    "absolute z-0 top-[40%] sm:top-[30%] left-[10%] h-[60%] w-[80%] bg-center bg-no-repeat bg-cover [opacity:0.4]",
-    props.sectionBackgroundName,
-  ]);
-
-  const baseCodeSnippetClass =
-    "hidden flex-col w-full h-full justify-between items-center gap-7 max-w-3xl";
-
-  const horizontalLayoutCodeSnippet = cc([baseCodeSnippetClass, "xl:flex"]);
-  const verticalLayoutCodeSnippet = cc([
-    baseCodeSnippetClass,
-    "md:flex xl:hidden",
-  ]);
-  const rightBasedCodeSnippet = cc([baseCodeSnippetClass, "md:flex"]);
-
-  const getCodeSnippet = (layoutClass: string) => {
-    return props.integrationsPanel ? (
-      <div className={layoutClass}>
-        <CodeSnippet codeSnippet={props.codeSnippet} />
-        <div className="inline-flex items-center gap-6">
-          {props.integrationsPanel}
-        </div>
-      </div>
-    ) : (
-      <div className={layoutClass}>
-        <CodeSnippet codeSnippet={props.codeSnippet} />
-      </div>
-    );
-  };
-
   return (
     <section className="relative w-full h-full py-10 px-5 flex flex-col items-center justify-center bg-white sm:bg-[radial-gradient(#DEDEDEB2_1px,transparent_1px)] sm:[background-size:24px_24px] md:py-24 sm:px-0">
       <div className="max-w-full flex flex-col items-center sm:mx-auto sm:px-6 sm:gap-3">
@@ -67,88 +35,42 @@ export function ReadySection(props: ReadySectionProps) {
             {props.sectionDescription}
           </p>
         </div>
-        {props.isCodeOnTheLeft ? (
-          <div className={mediaDiv}>
-            <div className={backgroundDiv} />
-            {getCodeSnippet(horizontalLayoutCodeSnippet)}
-            <Image
-              className="z-10 w-full h-full max-w-[34rem] xl:w-1/3 xl:h-auto"
-              src={props.imageSpecifics.source}
-              alt="chat image"
-              width={props.imageSpecifics.tagWidth}
-              height={props.imageSpecifics.tagHeight}
-              quality={100}
-            />
-            {getCodeSnippet(verticalLayoutCodeSnippet)}
+
+        <div
+          className={cc([
+            "flex sm:flex-col xl:flex-row justify-center border-0 rounded-none items-center gap-12 w-full h-full xl:h-3/4 py-16 px-10 flex-shrink-0 sm:gap-10",
+            props.isCodeOnTheLeft ? "xl:flex-row-reverse" : "",
+          ])}
+        >
+          <div
+            className={cc([
+              "absolute z-0 top-[40%] sm:top-[30%] left-[10%] h-[60%] w-[80%] bg-center bg-no-repeat bg-cover [opacity:0.4]",
+              props.sectionBackgroundName,
+            ])}
+          />
+          <Image
+            className="z-10 w-full h-full max-w-[34rem] xl:w-1/3 xl:h-auto"
+            src={props.imageSpecifics.source}
+            alt="chat image"
+            width={props.imageSpecifics.tagWidth}
+            height={props.imageSpecifics.tagHeight}
+            quality={100}
+          />
+
+          <div
+            className={
+              "hidden flex-col w-full h-full justify-between items-center gap-7 max-w-3xl md:flex"
+            }
+          >
+            <CodeSnippet codeSnippet={props.codeSnippet} />
+            {props.integrationsPanel && (
+              <div className="inline-flex items-center gap-6">
+                {props.integrationsPanel}
+              </div>
+            )}
           </div>
-        ) : (
-          <div className={mediaDiv}>
-            <div className={backgroundDiv} />
-            <Image
-              className="z-10 w-full h-full max-w-[34rem] xl:w-1/3 xl:h-auto"
-              src={props.imageSpecifics.source}
-              alt="chat image"
-              width={props.imageSpecifics.tagWidth}
-              height={props.imageSpecifics.tagHeight}
-              quality={100}
-            />
-            {getCodeSnippet(rightBasedCodeSnippet)}
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-type CodeSnippetProps = {
-  codeSnippet: string;
-};
-
-function CodeSnippet(props: CodeSnippetProps) {
-  const nameGuardMethods = [
-    "getSecurePrimaryName",
-    "fakeEthNameCheck",
-    "inspectName",
-    "bulkInspectNames",
-  ];
-
-  const findMethods = () => {
-    const spans = highlight(props.codeSnippet).split("><");
-    const toReplace = /var\(--sh-identifier\)/gi;
-    const methodColor = "#2596be";
-
-    return spans
-      .map((elem: string) => {
-        for (const method of nameGuardMethods) {
-          if (elem.includes(method)) {
-            return elem.replace(toReplace, methodColor);
-          }
-        }
-        return elem;
-      })
-      .join("><");
-  };
-
-  const code = findMethods();
-
-  return (
-    <div className="hidden sm:block bg-black rounded-xl pb-4 max-w-full h-fit bg-gradient-to-b from-figma-black to-black z-10">
-      <div className="flex flex-col gap-2.5 px-2.5 py-3">
-        <div className="flex justify-start gap-2">
-          <div className="rounded-full w-3 h-3 bg-[#434446]"></div>
-          <div className="rounded-full w-3 h-3 bg-[#434446]"></div>
-          <div className="rounded-full w-3 h-3 bg-[#434446]"></div>
         </div>
       </div>
-      <hr className="border-code-gray" />
-      <div className="py-4 px-5 max-w-full">
-        <pre className="w-full overflow-x-auto pb-4">
-          <code
-            dangerouslySetInnerHTML={{ __html: code }}
-            className="inline-block [overflow-wrap:break-word]"
-          />
-        </pre>
-      </div>
-    </div>
+    </section>
   );
 }
